requirements.txt
pypdf>=3.0
chromadb>=0.4
sentence-transformers>=2.2
openai>=1.0
python-dotenv>=1.0
ollama>=0.1
fastapi>=0.104
uvicorn>=0.24
python-multipart>=0.0.6
